moved the Browser component out
